+++ CHANGELOG.md
@@ -5,13 +5,16 @@
 
 ## What's Changed
 * fix: fixed naming of local file picker
+* feat: made userInterface into object
+* feat: added password strength measurement
+* fix: fixed spelling
 
 ## 0.1.3
 
 ## What's Changed
-* feat: implemeted open plain file FileHandler and its tests
-* feat: implemeted ceasar cypher as a placeholder (will be replaced in the future with acual algo)
-* feat: implemeted placeholder unit tests
+* feat: implemented open plain file FileHandler and its tests
+* feat: implemented cesar cypher as a placeholder (will be replaced in the future with actual algo)
+* feat: implemented placeholder unit tests
 * chore: added pyproject.toml
 * feat: Added local file picker based on https://github.com/zauberzeug/nicegui/blob/main/examples/local_file_picker/local_file_picker.py